web/requirements.txt
boto3==1.9.23
botocore==1.12.23
dataclasses==0.6
Django==2.1
django-bootstrap4==0.0.7
django-jquery==3.1.0
django-mathfilters==0.4.0
django-storages==1.7.1
djongo==1.2.30
dnspython==1.15.0
docutils==0.14
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pymongo==3.7.2
python-dateutil==2.7.3
python-decouple==3.1
python-memcached==1.59
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
whitenoise==4.1
django-heroku==0.3.1
dj-database-url==0.5.0
